+++ app/Http/Controllers/PertanyaanKuisController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\PertanyaanKuis;
+use App\Http\Resources\PertanyaanKuisResource;
+use Illuminate\Support\Facades\Auth;
 
 class PertanyaanKuisController extends Controller
 {
@@ -11,16 +13,26 @@ class PertanyaanKuisController extends Controller
     public function index()
     {
         $pertanyaanKuis = PertanyaanKuis::with('kuis')->get();
-        return response()->json([
-            'success' => true,
-            'message' => 'List of all pertanyaan kuis',
-            'data' => $pertanyaanKuis
-        ]);
+        return response()->json(
+            [
+                'status' => 'success',
+                'message' => 'Daftar pertanyaan berhasil di ambil',
+                'data' => PertanyaanKuisResource::collection($pertanyaanKuis)
+            ],
+            200
+        );
     }
 
     // Menyimpan data pertanyaan kuis baru.
     public function store(Request $request)
     {
+        // Check if user is authenticated and has the role of "guru"
+        if (Auth::user()->peran !== 'guru') {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 401);
+        }
+
         $validatedData = $request->validate([
             'id_kuis' => 'required|exists:kuis,id_kuis',
             'pertanyaan' => 'required|string',
@@ -28,11 +40,14 @@ class PertanyaanKuisController extends Controller
         ]);
 
         $pertanyaanKuis = PertanyaanKuis::create($validatedData);
-        return response()->json([
-            'success' => true,
-            'message' => 'Pertanyaan kuis created successfully',
-            'data' => $pertanyaanKuis
-        ], 201);
+        return response()->json(
+            [
+                'status' => 'success',
+                'message' => 'pertanyaan berhasil dibuat',
+                'data' => new PertanyaanKuisResource($pertanyaanKuis)
+            ],
+            201
+        );
     }
 
     // Menampilkan data pertanyaan kuis tertentu.
@@ -46,15 +61,19 @@ class PertanyaanKuisController extends Controller
             ], 404);
         }
 
-        return response()->json([
-            'success' => true,
-            'data' => $pertanyaanKuis
-        ]);
+        return new PertanyaanKuisResource($pertanyaanKuis);
     }
 
     // Mengupdate data pertanyaan kuis.
     public function update(Request $request, $id)
     {
+        // Check if user is authenticated and has the role of "guru"
+        if (Auth::user()->peran !== 'guru') {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 401);
+        }
+
         $validatedData = $request->validate([
             'id_kuis' => 'required|exists:kuis,id_kuis',
             'pertanyaan' => 'required|string',
@@ -71,27 +90,36 @@ class PertanyaanKuisController extends Controller
 
         $pertanyaanKuis->update($validatedData);
         return response()->json([
-            'success' => true,
-            'message' => 'Pertanyaan kuis updated successfully',
-            'data' => $pertanyaanKuis
+            'status' => 'success',
+            'message' => 'pertanyaan berhasil diperbarui',
+            'data' => new PertanyaanKuisResource($pertanyaanKuis)
         ], 200);
     }
 
     // Menghapus data pertanyaan kuis.
     public function destroy($id)
     {
+        // Check if user is authenticated and has the role of "guru"
+        if (Auth::user()->peran !== 'guru') {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 401);
+        }
+        
         $pertanyaanKuis = PertanyaanKuis::find($id);
         if (!$pertanyaanKuis) {
             return response()->json([
-                'success' => false,
-                'message' => 'Pertanyaan kuis not found'
+                'status' => 'error',
+                'message' => 'pertanyaanKuis tidak ditemukan'
             ], 404);
         }
 
+        // Delete the berita
         $pertanyaanKuis->delete();
+
         return response()->json([
-            'success' => true,
-            'message' => 'Pertanyaan kuis deleted successfully'
-        ], 204);
+            'status' => 'success',
+            'message' => 'pertanyaanKuis berhasil dihapus'
+        ], 200);
     }
 }
